Test submimssion commit: id=23642280, title=피보나치 수 2

+++ baekjoon/2748.cpp
@@ -10,7 +10,7 @@
 
 using namespace std;
 
-int dp[91] = { 0, 1 };
+long long dp[91] = { 0, 1 };
 
 int main() {
 	ios_base::sync_with_stdio(false);
